Moved the verification problems from the testMesh unit test to their own unit test file
